requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
dj-database-url==2.3.0
Django==5.2
django-cloudinary-storage==0.3.0
django-tinymce==4.1.0
dotenv==0.9.9
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
whitenoise==6.9.0
